Refactor test support to remove XCTest dependency.
